+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+Version 0.3.0 *(15-08-2016)*
+----------------------------
+
+* Face Detector initialization is moved to background thread.
+* CenterFaceCrop is changed to **FaceCenterCrop**
+* Multiple faces in the image are being considered.
+* Bug fixes
+
+**Library dependencies updated:**
+```java
+com.google.android.gms:play-services-vision:9.4.0
+com.github.bumptech.glide:glide:3.7.0
+```
+
 Version 0.2.1 *(24-07-2016)*
 --------------------------